some new glsl shaders, refactoring codes and get rid of most of the warnings in Xcode

+++ src/demi/misc/Str.cpp
@@ -1289,7 +1289,7 @@ namespace Demi
     DiString DiString::ExtractBaseName() const
     {
         DiString basename;
-        size_t i = rfind(".");
+        int i = rfind(".");
         if (i == npos)
         {
             basename = *this;
